Initial serverThread graceful shutdown
threadMain of serverThread runs while terminate() has not been called on event

+++ server.py
@@ -11,19 +11,19 @@ class serverThread(Thread):
 
     def threadMain(self):
         # Handles each incoming connection request
-        while True:
+        while not self.stop.wait(1):
             try: 
                 self._running = True
                 client, client_address = server.accept()
-                #print("%s:%s has connected." % client_address)
+                print("%s:%s has connected." % client_address)
                 client.send(bytes("Please enter your username", "utf8"))
                 addresses[client] = client_address
                 Thread(target=socketThread, args=(client,)).start()
             except:
-                print("Failed to connect to client")
+                print("\nFailed to connect to client")
             finally:
                 self._running = False
-    
+                
     def terminate(self):
         self.stop.set()
 
@@ -33,15 +33,18 @@ def socketThread(client):
 
     name = client.recv(buffer).decode("utf8")
     welcome = 'Type {disconnect} to exit.'
-    client.send(bytes(welcome, "utf8"))
+    try:
+        client.send(bytes(welcome, "utf8"))
+    except:
+        print('\nCould not send welcome message')
     
     handle = {'socket' : client, 'room' : ''}
     clients[name] = handle
-    roomPrompt = 'Please enter the chatroom you wish to join'
+    roomPrompt = '\nPlease enter the chatroom you wish to join'
     try:
         client.send(bytes(roomPrompt, "utf8"))
     except:
-        print("Failed to send room prompt message to client")
+        print("\nFailed to send room prompt message to client")
     
 
     while True:
@@ -56,9 +59,10 @@ def socketThread(client):
         else:
             try:
                 client.send(bytes("{disconnect}", "utf8"))
-                broadcast(bytes("%s has left the chat." % name, "utf8"), room)
             except:
                 pass
+            finally:
+                broadcast(bytes("%s has left the chat." % name, "utf8"), room)
             break
             
 
@@ -67,12 +71,11 @@ def broadcast(msg, room, prefix=""):  # prefix is for name identification.
     """Broadcasts a message to all the clients."""
 
     for user in clients.values():
-        print(user['socket'])
         if user['room'] == room:
             try:
                 user['socket'].send(bytes(prefix, "utf8")+msg)
             except:
-                pass
+                print("Client " +str(user['socket'])+ " has disconnected")
 
         
 clients = {}
@@ -81,16 +84,22 @@ addresses = {}
 host = ''
 port = 4200
 buffer = 1024
-ADDR = (host, port)
+address = (host, port)
 
 server = socket(AF_INET, SOCK_STREAM)
-server.bind(ADDR)
+server.bind(address)
 
 if __name__ == "__main__":
     server.listen(5)
-    print("Waiting for connection...")
+    print("Waiting for connections...\n")
     server_thread = serverThread()
     server_thread.start()
-    server_thread.terminate()
-    server_thread.join()
+    
+    
+    while True:
+        terminate = input("Enter 'shutdown' to terminate server\n")
+        if terminate == 'shutdown':
+            server_thread.terminate()
+            server_thread.join()
+            break
     server.close()
